+++ CHANGES.rst
@@ -190,3 +190,4 @@ v1.1.7
 v1.1.8[unreleased]
 -----------
 * Add unspecified gender option
+* Stop requiring email on password recovery